# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AddToCartQuantityRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
  models.AddToCartRequest:
    properties:
      product_id:
        type: string
    required:
    - product_id
    type: object
  models.CategoryRequest:
    properties:
      id:
        type: string
      image_url:
        type: string
      is_delete:
        type: boolean
      name:
        type: string
      uid:
        type: string
    type: object
  models.CreateUserRequest:
    properties:
      email:
        type: string
      is_owner:
        type: boolean
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.OtpVerifyRequest:
    properties:
      otp:
        type: string
      uid:
        type: string
    required:
    - otp
    - uid
    type: object
  models.ProductRequest:
    properties:
      des:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: integer
    required:
    - name
    - price
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_owner:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  utils.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact: {}
  description: API documentation for Amar Dokan.
  title: Amar Dokan API
  version: "1.0"
paths:
  /add-to-cart:
    get:
      consumes:
      - application/json
      description: Get all cart items for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get cart items
      tags:
      - Add To Cart
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated user's cart
      parameters:
      - description: Add to cart payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Add to cart successfully
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Payload is not valid
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Add product to cart
      tags:
      - Add To Cart
  /add-to-cart/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific item from the authenticated user's cart
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete cart item
      tags:
      - Add To Cart
    put:
      consumes:
      - application/json
      description: Update quantity for a specific cart item
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      - description: Quantity payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.AddToCartQuantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: quantity Update Successfully
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Add To Cart
  /category:
    get:
      consumes:
      - application/json
      description: Get all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category create Successful
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Categories
  /products:
    get:
      consumes:
      - application/json
      description: Get all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Note Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product update successfully
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Products
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Account created. OTP has been sent to your email.
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Create new user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: authenticate user and get token
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login Successful
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Login user
      tags:
      - Users
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
  /users/refresh-token:
    get:
      consumes:
      - application/json
      description: Generate a new access token using the Authorization bearer token
      parameters:
      - description: Bearer refresh token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Access Token Genaret Successful
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Refresh access token
      tags:
      - Users
  /users/verification:
    post:
      consumes:
      - application/json
      description: Verify a newly registered user's OTP and return an access token
      parameters:
      - description: OTP verification payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.OtpVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Otp Verification Successful
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Verify OTP
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"